python: handling for API edge cases
